Add new color scheme, add panning, fix zooming, add high zoom warning
There is now a new color scheme for coloring the fractal that changes as the fractal is zoomed in more. Holding shift and clicking will pan the fractal to the location of clicking. Click zooming used to not be able to handle zooming in and out sequentially and now that's fixed. There is now a warning for when javascript floating point numbers can't represent the detail of the fractal anymore.